+++ middleman-core/lib/middleman-core/util/data.rb
@@ -45,13 +45,14 @@ module Middleman
       # @param [String] path
       # @return [Array<Hash, String>]
       Contract IsA['Middleman::SourceFile'], Maybe[Symbol] => [Hash, Maybe[String]]
-      def parse(file, frontmatter_delims, known_type = nil)
-        full_path = file[:full_path]
-        return [{}, nil] if ::Middleman::Util.binary?(full_path) || file[:types].include?(:binary)
+      def parse(source_file, frontmatter_delims, known_type = nil)
+        full_path = source_file[:full_path]
+
+        return [{}, nil] if ::Middleman::Util.binary?(full_path) || source_file[:types].include?(:binary)
 
         # Avoid weird race condition when a file is renamed
         begin
-          content = file.read
+          content = source_file.read
         rescue EOFError, IOError, ::Errno::ENOENT
           return [{}, nil]
         end